allow pdef input in run_adam.py
