+++ run_adam.outfile
@@ -1,16 +1,2 @@
 Welcome to ADAM!
 You have just started running, this file will print updates for each step
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-ADAM was interrupted, retarting from step 598
-597, 0.538021
-598, 0.538021
